read_pdf_invoices/invoice_read_lambda/requirements-lambda.txt
attrs==23.1.0
cattrs==23.1.2
colored
constructs==10.3.0
dateparser==1.2.0
exceptiongroup==1.1.3
importlib-resources==6.1.1
iniconfig==2.0.0
invoice2data==0.4.4
jmespath==1.0.1
jsii==1.91.0
numpy==1.26.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pluggy==1.3.0
publication==0.0.3
pytest==7.4.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
six==1.16.0
syrupy
tomli==2.0.1
typeguard==2.13.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3<2
